all at once
